# internal/web/api.go
// Package web also hosts the small HTTP API skills use while they run.
// The surface mirrors openapi.yaml at the repo root: list scans, read a
// scan, list skills, enqueue a skill scan, fetch a repository summary.
// Requests authenticate with a per-scan bearer token that the worker
// stages into the workspace's context.json.
package web

import (
	"context"
	"crypto/rand"
	"encoding/hex"
	"encoding/json"
	"errors"
	"fmt"
	"io"
	"net/http"
	"slices"
	"strconv"
	"strings"

	"gorm.io/gorm"

	"scrutineer/internal/db"
	"scrutineer/internal/worker"
)

const apiPrefix = "/api"

// maxAgentAPIOpenScansPerRepository bounds model-backed work a compromised
// scan token can enqueue. Normal orchestration stays well below this ceiling.
const maxAgentAPIOpenScansPerRepository = 16

// NewAPIToken returns a 32-byte hex token suitable for bearer auth.
func NewAPIToken() string {
	var b [32]byte
	_, _ = rand.Read(b[:])
	return hex.EncodeToString(b[:])
}

type apiCtxKey struct{}

// scanBlobColumns are the wide columns on a scan row. A running scan's log
// grows for the length of the run, so endpoints serving only a scan's identity
// or summary omit them; apiGetScan, which returns the report and log, does not.
var scanBlobColumns = []string{"Log", "Prompt", "Report", "RefusalAudit", "ImportPayload"}

// authScanOmitColumns additionally drops the scoping blobs, which nothing
// reachable from scanFromRequest reads.
var authScanOmitColumns = slices.Concat(scanBlobColumns,
	[]string{"FocusArea", "DiffStats", "Coverage"})

// repositoryBlobColumns are the wide columns on a repository row, dominated by
// the cached ecosyste.ms payloads.
var repositoryBlobColumns = []string{
	"Metadata", "ThreatModel", "ScanConfig",
	"EcosystemsRepoData", "EcosystemsPackagesData", "EcosystemsAdvisoriesData",
	"EcosystemsCommitsData", "EcosystemsIssuesData", "EcosystemsDependentsData",
}

// apiAuth validates bearer tokens against the currently running scan rows
// and puts the scan on the request context so handlers can apply the
// "skills only touch their own repo" rule.
func (s *Server) apiAuth(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		token := bearer(r.Header.Get("Authorization"))
		if token == "" {
			writeAPIError(w, http.StatusUnauthorized, "missing bearer token")
			return
		}
		var scan db.Scan
		if err := s.DB.Omit(authScanOmitColumns...).
			Where("api_token = ? AND status = ?", token, db.ScanRunning).
			First(&scan).Error; err != nil {
			if errors.Is(err, gorm.ErrRecordNotFound) {
				writeAPIError(w, http.StatusUnauthorized, "token invalid or scan not running")
				return
			}
			writeAPIError(w, http.StatusInternalServerError, err.Error())
			return
		}
		ctx := context.WithValue(r.Context(), apiCtxKey{}, &scan)
		r.Body = http.MaxBytesReader(w, r.Body, apiMaxBody)
		next.ServeHTTP(w, r.WithContext(ctx))
	})
}

const apiMaxBody = 1 << 20

//nolint:ireturn // T is a concrete struct at every call site, not an interface
func decodeAPIBody[T any](w http.ResponseWriter, r *http.Request, errorMessage string) (T, bool) {
	var body T
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		writeAPIError(w, http.StatusBadRequest, errorMessage)
		return body, false
	}
	return body, true
}

func decodeOptionalAPIBody(w http.ResponseWriter, r *http.Request, dst any) bool {
	if err := json.NewDecoder(r.Body).Decode(dst); err != nil {
		if errors.Is(err, io.EOF) {
			return true
		}
		writeAPIError(w, http.StatusBadRequest, "invalid JSON request body")
		return false
	}
	return true
}

func bearer(h string) string {
	const prefix = "Bearer "
	if strings.HasPrefix(h, prefix) {
		return strings.TrimSpace(h[len(prefix):])
	}
	return ""
}

// scanFromRequest pulls the authenticated scan off the request context.
func scanFromRequest(r *http.Request) *db.Scan {
	if v, ok := r.Context().Value(apiCtxKey{}).(*db.Scan); ok {
		return v
	}
	return nil
}

// scanOwnsRepo enforces the rule that a scan's API token only grants access
// to the repository it was issued against.
func (s *Server) scanOwnsRepo(r *http.Request, repoID uint) bool {
	sc := scanFromRequest(r)
	return sc != nil && sc.RepositoryID == repoID
}

func (s *Server) apiHandler() http.Handler {
	mux := http.NewServeMux()
	mux.HandleFunc("GET /repositories/{id}", s.apiGetRepository)
	mux.HandleFunc("PATCH /repositories/{id}", s.apiPatchRepository)
	mux.HandleFunc("GET /repositories/{id}/scans", s.apiListScans)
	mux.HandleFunc("GET /repositories/{id}/maintainers", s.apiListMaintainers)
	mux.HandleFunc("GET /repositories/{id}/packages", s.apiListPackages)
	mux.HandleFunc("GET /repositories/{id}/alternatives", s.apiListPackageAlternatives)
	mux.HandleFunc("GET /repositories/{id}/advisories", s.apiListAdvisories)
	mux.HandleFunc("GET /repositories/{id}/dependents", s.apiListDependents)
	mux.HandleFunc("GET /repositories/{id}/ecosystems/{source}/raw", s.apiGetEcosystemsRaw)
	mux.HandleFunc("GET /repositories/{id}/expected", s.apiListExpectedFindings)
	mux.HandleFunc("GET /repositories/{id}/dependencies", s.apiListDependencies)
	mux.HandleFunc("GET /repositories/{id}/findings", s.apiListFindings)
	mux.HandleFunc("POST /repositories/{id}/findings", s.apiStreamFinding)
	mux.HandleFunc("GET /repositories/{id}/dependency-findings", s.apiListDependencyFindings)
	mux.HandleFunc("POST /repositories/{id}/skills/{name}/run", s.apiRunSkill)
	mux.HandleFunc("POST /findings/{id}/skills/{name}/run", s.apiRunFindingSkill)
	mux.HandleFunc("GET /scans/{id}", s.apiGetScan)
	mux.HandleFunc("POST /scans/{id}/validate-report", s.apiValidateReport)
	mux.HandleFunc("GET /findings/{id}", s.apiGetFinding)
	mux.HandleFunc("PATCH /findings/{id}", s.apiPatchFinding)
	mux.HandleFunc("GET /findings/{id}/notes", s.apiListFindingNotes)
	mux.HandleFunc("POST /findings/{id}/notes", s.apiAddFindingNote)
	mux.HandleFunc("GET /findings/{id}/reviews", s.apiListFindingReviews)
	// /audit/queue and /audit/metrics are intentionally on the host-only
	// /api/v1 export mux, not here: they return findings across every
	// repository on the instance, so a scan token issued for one repo
	// must not be able to read them (#454).
	mux.HandleFunc("GET /findings/{id}/communications", s.apiListFindingCommunications)
	mux.HandleFunc("POST /findings/{id}/communications", s.apiAddFindingCommunication)
	mux.HandleFunc("GET /findings/{id}/references", s.apiListFindingReferences)
	mux.HandleFunc("POST /findings/{id}/references", s.apiAddFindingReference)
	mux.HandleFunc("PUT /findings/{id}/labels", s.apiSetFindingLabels)
	mux.HandleFunc("GET /findings/{id}/history", s.apiListFindingHistory)
	mux.HandleFunc("GET /skills", s.apiListSkills)
	mux.HandleFunc("GET /cnas", s.apiListCNAs)
	return http.StripPrefix(apiPrefix, s.apiAuth(mux))
}

func (s *Server) apiGetRepository(w http.ResponseWriter, r *http.Request) {
	id, _ := strconv.Atoi(r.PathValue("id"))
	if !s.scanOwnsRepo(r, uint(id)) {
		writeAPIError(w, http.StatusForbidden, "scan may only read its own repository")
		return
	}
	var repo db.Repository
	if err := s.DB.Omit(repositoryBlobColumns...).First(&repo, id).Error; err != nil {
		writeAPIError(w, http.StatusNotFound, "repository not found")
		return
	}
	writeJSON(w, http.StatusOK, map[string]any{
		"id":              repo.ID,
		"url":             repo.URL,
		"name":            repo.Name,
		"full_name":       repo.FullName,
		"default_branch":  repo.DefaultBranch,
		"html_url":        repo.HTMLURL,
		"stars":           repo.Stars,
		"forks":           repo.Forks,
		"archived":        repo.Archived,
		"languages":       repo.Languages,
		"license":         repo.License,
		"fork":            repo.Fork,
		"posture":         repo.Posture,
		"posture_summary": repo.PostureSummary,
		"health":          repo.Health,
	})
}

// ecosystemsRawColumns maps the {source} path segment of the diagnostic
// endpoint to the Repository column holding that source's cached payload.
var ecosystemsRawColumns = map[string]string{
	"repo":       "ecosystems_repo_data",
	"packages":   "ecosystems_packages_data",
	"advisories": "ecosystems_advisories_data",
	"commits":    "ecosystems_commits_data",
	"issues":     "ecosystems_issues_data",
	"dependents": "ecosystems_dependents_data",
}

// apiGetEcosystemsRaw returns the verbatim cached ecosyste.ms payload for one
// source: an operator/debug escape hatch, and a skill fallback when a
// digested endpoint does not cover an edge case. 404 when nothing is cached
// (the skill then falls back to WebFetch, which is also the steady state under
// `ecosystems_enrichment: false`, where no source is ever cached); 400 for an
// unknown source.
func (s *Server) apiGetEcosystemsRaw(w http.ResponseWriter, r *http.Request) {
	id, _ := strconv.Atoi(r.PathValue("id"))
	if !s.scanOwnsRepo(r, uint(id)) {
		writeAPIError(w, http.StatusForbidden, "scan may only read its own repository")
		return
	}
	column, ok := ecosystemsRawColumns[r.PathValue("source")]
	if !ok {
		writeAPIError(w, http.StatusBadRequest, "unknown ecosystems source")
		return
	}
	var payload string
	if err := s.DB.Model(&db.Repository{}).Where("id = ?", id).Pluck(column, &payload).Error; err != nil {
		writeAPIError(w, http.StatusInternalServerError, err.Error())
		return
	}
	if payload == "" {
		writeAPIError(w, http.StatusNotFound, "no cached payload for source")
		return
	}
	w.Header().Set("Content-Type", "application/json; charset=utf-8")
	_, _ = w.Write([]byte(payload))
}

// apiPatchRepository lets a skill write back fields it derived for the
// repository. Currently only `fork` (the staging fork's owner/name) is
// accepted; everything else on Repository is owned by the metadata job.
func (s *Server) apiPatchRepository(w http.ResponseWriter, r *http.Request) {
	id, _ := strconv.Atoi(r.PathValue("id"))
	if !s.scanOwnsRepo(r, uint(id)) {
		writeAPIError(w, http.StatusForbidden, "scan may only edit its own repository")
		return
	}
	body, ok := decodeAPIBody[struct {
		Fork *string `json:"fork"`
	}](w, r, "body must be JSON")
	if !ok {
		return
	}
	if body.Fork == nil {
		writeAPIError(w, http.StatusUnprocessableEntity, "no writable fields in body")
		return
	}
	if err := s.DB.Model(&db.Repository{}).Where("id = ?", id).Update("fork", *body.Fork).Error; err != nil {
		writeAPIError(w, http.StatusInternalServerError, err.Error())
		return
	}
	w.WriteHeader(http.StatusNoContent)
}

func (s *Server) apiListScans(w http.ResponseWriter, r *http.Request) {
	id, _ := strconv.Atoi(r.PathValue("id"))
	if !s.scanOwnsRepo(r, uint(id)) {
		writeAPIError(w, http.StatusForbidden, "scan may only list scans on its own repository")
		return
	}
	q := s.DB.Where("repository_id = ?", id).Order("id desc")
	if status := r.URL.Query().Get(statusKey); status != "" {
		q = q.Where("status = ?", status)
	}
	if skill := r.URL.Query().Get("skill"); skill != "" {
		q = q.Where("skill_name = ?", skill)
	}
	var rows []db.Scan
	q.Omit(scanBlobColumns...).Find(&rows)
	out := make([]map[string]any, 0, len(rows))
	for _, sc := range rows {
		out = append(out, scanSummary(sc))
	}
	writeJSON(w, http.StatusOK, out)
}

func (s *Server) apiGetScan(w http.ResponseWriter, r *http.Request) {
	id, _ := strconv.Atoi(r.PathValue("id"))
	var sc db.Scan
	if err := s.DB.First(&sc, id).Error; err != nil {
		writeAPIError(w, http.StatusNotFound, "scan not found")
		return
	}
	if !s.scanOwnsRepo(r, sc.RepositoryID) {
		writeAPIError(w, http.StatusForbidden, "scan may only read scans on its own repository")
		return
	}
	summary := scanSummary(sc)
	summary["report"] = sc.Report
	summary["refusal_audit"] = sc.RefusalAudit
	summary["log"] = sc.Log
	writeJSON(w, http.StatusOK, summary)
}

// apiValidateReport lets a running skill check a candidate report.json against
// its skill's schema without installing a JSON Schema library inside the runner
// container. The request body is the candidate report; the response is
// {"valid":true} or {"valid":false,"errors":"..."} using the exact same
// validator (worker.ValidateSkillReport) the harness runs after the scan, so
// an in-container pass guarantees the harness will not send a repair prompt.
//
// The body is already capped at apiMaxBody (1 MB) by apiAuth's MaxBytesReader;
// an oversized report is rejected rather than silently truncated.
func (s *Server) apiValidateReport(w http.ResponseWriter, r *http.Request) {
	id, _ := strconv.Atoi(r.PathValue("id"))
	scan := scanFromRequest(r)
	if scan == nil || scan.ID != uint(id) {
		writeAPIError(w, http.StatusForbidden, "scan may only validate its own report")
		return
	}
	if scan.SkillID == nil {
		writeAPIError(w, http.StatusBadRequest, "scan has no skill")
		return
	}
	var skill db.Skill
	if err := s.DB.First(&skill, *scan.SkillID).Error; err != nil {
		writeAPIError(w, http.StatusNotFound, "skill not found")
		return
	}
	if skill.SchemaJSON == "" {
		writeJSON(w, http.StatusOK, map[string]any{"valid": true, "note": "skill has no schema"})
		return
	}
	body, err := io.ReadAll(r.Body)
	if err != nil {
		writeAPIError(w, http.StatusRequestEntityTooLarge, "could not read body (max 1 MB)")
		return
	}
	if detail := worker.ValidateSkillReport(skill.Name, skill.SchemaJSON, string(body)); detail != "" {
		writeJSON(w, http.StatusOK, map[string]any{"valid": false, "errors": detail})
		return
	}
	writeJSON(w, http.StatusOK, map[string]any{"valid": true})
}

func (s *Server) apiRunSkill(w http.ResponseWriter, r *http.Request) {
	id, _ := strconv.Atoi(r.PathValue("id"))
	name := r.PathValue("name")
	if !s.scanOwnsRepo(r, uint(id)) {
		writeAPIError(w, http.StatusForbidden, "scan may only trigger skills on its own repository")
		return
	}
	var skill db.Skill
	if err := s.DB.Where("name = ? AND active = ?", name, true).First(&skill).Error; err != nil {
		writeAPIError(w, http.StatusNotFound, "skill not found or inactive")
		return
	}
	var body struct {
		Model          string `json:"model"`
		Ref            string `json:"ref"`
		Profile        string `json:"profile"`
		RescanMode     string `json:"rescan_mode"`
		SubPath        string `json:"sub_path"`
		BaselineScanID *uint  `json:"baseline_scan_id"`
	}
	if !decodeOptionalAPIBody(w, r, &body) {
		return
	}
	if body.Profile != "" && !worker.KnownProfile(body.Profile) {
		writeAPIError(w, http.StatusBadRequest, "unknown profile")
		return
	}
	// sub_path scopes this run to a monorepo sub-package; triage forwards it to
	// each pipeline child so the whole scan set stays scoped. Validated here so
	// a traversal attempt is rejected before it can reach workspace staging.
	subPath, err := worker.CleanSubPath(body.SubPath)
	if err != nil {
		writeAPIError(w, http.StatusBadRequest, err.Error())
		return
	}
	s.agentEnqueueMu.Lock()
	defer s.agentEnqueueMu.Unlock()
	if s.hasOpenRepoScopedScan(uint(id), skill.ID, subPath) {
		writeAPIError(w, http.StatusConflict, "equivalent scan already queued or running")
		return
	}
	if !s.agentAPIRepoHasCapacity(w, uint(id)) {
		return
	}
	scanID, err := s.enqueueSkillWith(r.Context(), uint(id), skill.ID, ScanOpts{
		Model:          body.Model,
		Ref:            body.Ref,
		Profile:        body.Profile,
		RescanMode:     body.RescanMode,
		SubPath:        subPath,
		DiffBaseScanID: body.BaselineScanID,
	})
	if err != nil {
		if errors.Is(err, ErrSkillRequiresRemote) {
			writeAPIError(w, http.StatusNotFound, err.Error())
			return
		}
		if errors.Is(err, ErrRepoFederationOptOut) {
			writeAPIError(w, http.StatusConflict, err.Error())
			return
		}
		if errors.Is(err, ErrSkillProfileMismatch) {
			writeAPIError(w, http.StatusBadRequest, err.Error())
			return
		}
		if errors.Is(err, ErrInvalidRef) {
			writeAPIError(w, http.StatusBadRequest, err.Error())
			return
		}
		if errors.Is(err, ErrInvalidRescanMode) {
			writeAPIError(w, http.StatusBadRequest, err.Error())
			return
		}
		writeAPIError(w, http.StatusInternalServerError, err.Error())
		return
	}
	var sc db.Scan
	s.DB.First(&sc, scanID)
	writeJSON(w, http.StatusCreated, scanSummary(sc))
}

// apiRunFindingSkill enqueues a finding-scoped skill (verify, patch,
// reattack, disclose). The authenticated scan must be on the same repository that
// owns the finding.
func (s *Server) apiRunFindingSkill(w http.ResponseWriter, r *http.Request) {
	id, _ := strconv.Atoi(r.PathValue("id"))
	name := r.PathValue("name")
	repoID, ok := s.findingRepoID(uint(id))
	if !ok {
		writeAPIError(w, http.StatusNotFound, "finding not found")
		return
	}
	if !s.scanOwnsRepo(r, repoID) {
		writeAPIError(w, http.StatusForbidden, "scan may only trigger skills on its own repository")
		return
	}
	var skill db.Skill
	if err := s.DB.Where("name = ? AND active = ?", name, true).First(&skill).Error; err != nil {
		writeAPIError(w, http.StatusNotFound, "skill not found or inactive")
		return
	}
	var body struct {
		Model string `json:"model"`
	}
	if !decodeOptionalAPIBody(w, r, &body) {
		return
	}
	s.agentEnqueueMu.Lock()
	defer s.agentEnqueueMu.Unlock()
	opts, err := s.findingSkillScanOpts(uint(id), name, body.Model)
	if err != nil {
		writeAPIError(w, http.StatusPreconditionFailed, err.Error())
		return
	}
	if opts.RemediationAttemptID == nil && s.hasOpenFindingScopedScan(uint(id), skill.ID) {
		writeAPIError(w, http.StatusConflict, "equivalent scan already queued or running")
		return
	}
	if opts.RemediationAttemptID != nil && s.hasOpenScan(
		"finding_id = ? AND skill_id = ? AND remediation_attempt_id = ?",
		uint(id), skill.ID, *opts.RemediationAttemptID) {
		writeAPIError(w, http.StatusConflict, "equivalent scan already queued or running")
		return
	}
	if !s.agentAPIRepoHasCapacity(w, repoID) {
		return
	}
	opts.FindingID = new(uint(id))
	scanID, err := s.enqueueSkillWith(r.Context(), repoID, skill.ID, opts)
	if err != nil {
		if errors.Is(err, ErrSkillRequiresRemote) {
			writeAPIError(w, http.StatusNotFound, err.Error())
			return
		}
		if errors.Is(err, ErrRepoFederationOptOut) || errors.Is(err, ErrFederationClaimPending) {
			writeAPIError(w, http.StatusConflict, err.Error())
			return
		}
		writeAPIError(w, http.StatusInternalServerError, err.Error())
		return
	}
	var sc db.Scan
	s.DB.First(&sc, scanID)
	writeJSON(w, http.StatusCreated, scanSummary(sc))
}

// agentAPIRepoHasCapacity rejects scan-token enqueue requests once the target
// repository already has the maximum number of queued or running scans.
func (s *Server) agentAPIRepoHasCapacity(w http.ResponseWriter, repoID uint) bool {
	var open int64
	if err := s.DB.Model(&db.Scan{}).
		Where("repository_id = ? AND status IN ?", repoID, []db.ScanStatus{db.ScanQueued, db.ScanRunning}).
		Count(&open).Error; err != nil {
		writeAPIError(w, http.StatusInternalServerError, err.Error())
		return false
	}
	if open >= maxAgentAPIOpenScansPerRepository {
		writeAPIError(w, http.StatusTooManyRequests, "repository has too many queued or running scans")
		return false
	}
	return true
}

// findingRepoID reads the denormalized Finding.RepositoryID column. Used
// by the skill-facing handlers to enforce "scan can only touch findings
// on its own repository" without re-reading the entire Finding row.
func (s *Server) findingRepoID(findingID uint) (uint, bool) {
	var repoID uint
	row := s.DB.Model(&db.Finding{}).Select("repository_id").Where("id = ?", findingID).Row()
	if err := row.Scan(&repoID); err != nil || repoID == 0 {
		return 0, false
	}
	return repoID, true
}

// apiListCNAs returns the cached CVE Numbering Authority list. Global
// (not repo-scoped) since CNA scope is matched against repo metadata by
// the caller, not by scrutineer. Supports ?q= for a substring match
// across short_name, organization, and scope so a skill can narrow before
// reading prose.
func (s *Server) apiListCNAs(w http.ResponseWriter, r *http.Request) {
	q := s.DB.Order("short_name")
	if term := r.URL.Query().Get("q"); term != "" {
		like := "%" + term + "%"
		q = q.Where("short_name LIKE ? OR organization LIKE ? OR scope LIKE ?", like, like, like)
	}
	var rows []db.CNA
	q.Find(&rows)
	out := make([]map[string]any, 0, len(rows))
	for _, c := range rows {
		out = append(out, map[string]any{
			"short_name":   c.ShortName,
			"cna_id":       c.CNAID,
			"organization": c.Organization,
			"scope":        c.Scope,
			"email":        c.Email,
			"contact_url":  c.ContactURL,
			"policy_url":   c.PolicyURL,
			"advisory_url": c.AdvisoryURL,
			"root":         c.Root,
			"types":        c.Types,
		})
	}
	writeJSON(w, http.StatusOK, out)
}

func (s *Server) apiListSkills(w http.ResponseWriter, r *http.Request) {
	q := s.DB.Order("name")
	if v := r.URL.Query().Get("active"); v != "" {
		// A malformed value is a 400 rather than a silent fall-back to false,
		// so ?active=yes never quietly returns the inactive skills instead.
		active, err := strconv.ParseBool(v)
		if err != nil {
			writeAPIError(w, http.StatusBadRequest,
				fmt.Sprintf("active: must be true or false, got %q", v))
			return
		}
		q = q.Where("active = ?", active)
	}
	var rows []db.Skill
	q.Find(&rows)
	out := make([]map[string]any, 0, len(rows))
	for _, sk := range rows {
		out = append(out, map[string]any{
			"id":          sk.ID,
			"name":        sk.Name,
			"description": sk.Description,
			"output_kind": sk.OutputKind,
			"output_file": sk.OutputFile,
			"max_turns":   sk.MaxTurns,
			"model":       sk.Model,
			"version":     sk.Version,
			"active":      sk.Active,
		})
	}
	writeJSON(w, http.StatusOK, out)
}

func scanSummary(sc db.Scan) map[string]any {
	m := map[string]any{
		"id":                   sc.ID,
		"repository_id":        sc.RepositoryID,
		"kind":                 sc.Kind,
		statusKey:              string(sc.Status),
		"model":                sc.Model,
		"commit":               sc.Commit,
		"skill_name":           sc.SkillName,
		"skill_version":        sc.SkillVersion,
		"skill_schema_version": sc.SkillSchemaVersion,
		"started_at":           sc.StartedAt,
		"finished_at":          sc.FinishedAt,
		"max_turns_hit":        sc.MaxTurnsHit,
		errorKey:               sc.Error,
	}
	m["refusal_audit_warning"] = sc.RefusalAuditWarning
	if sc.Ref != "" {
		m["ref"] = sc.Ref
	}
	if sc.RescanMode != "" {
		m["rescan_mode"] = sc.RescanMode
	}
	if sc.DiffBaseScanID != nil {
		m["diff_base_scan_id"] = *sc.DiffBaseScanID
	}
	if sc.RemediationAttemptID != nil {
		m["remediation_attempt_id"] = *sc.RemediationAttemptID
	}
	if sc.DiffBaseCommit != "" {
		m["diff_base_commit"] = sc.DiffBaseCommit
	}
	if sc.DiffThreatModelScanID != nil {
		m["diff_threat_model_scan_id"] = *sc.DiffThreatModelScanID
	}
	if sc.DiffStats != "" {
		m["diff_stats"] = sc.DiffStats
	}
	if sc.Coverage != "" {
		m["coverage"] = sc.Coverage
	}
	return m
}

func writeJSON(w http.ResponseWriter, code int, v any) {
	w.Header().Set("Content-Type", "application/json; charset=utf-8")
	w.WriteHeader(code)
	_ = json.NewEncoder(w).Encode(v)
}

func writeAPIError(w http.ResponseWriter, code int, msg string) {
	writeJSON(w, code, map[string]string{errorKey: msg})
}