# openapi.yaml
openapi: 3.1.0
info:
  title: Scrutineer Skill API
  version: 0.4.0
  description: |
    A narrow HTTP surface that claude-code skills use while they run. The
    goal is enough to let a skill orchestrate other skills and read prior
    scan outputs without pushing that logic into scrutineer's Go code.

    Every scrutineer skill scan is issued a bearer token that is staged into
    the workspace context.json. Skills authenticate with that token. A token
    is only valid while its originating scan is running, and only grants
    access to the repository the scan was issued against.

    The `/v1/*` paths are an unauthenticated, localhost-only export surface
    intended for external scripts (jq, cron, CI dumps). They emit JSONL and
    rely on the same host-header check as the browser UI; see threatmodel.md.
servers:
  - url: http://127.0.0.1:8080/api
    description: Local scrutineer server
security:
  - bearerAuth: []
paths:
...
  /repositories/{id}:
    get:
      summary: Fetch a repository summary
      operationId: getRepository
      parameters:
        - $ref: '#/components/parameters/RepositoryID'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Repository' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      summary: Update skill-writable repository fields
      description: |
        Allows the authenticated scan to write repository fields that are
        produced by skills rather than owned by metadata refreshes. Currently
        only `fork` is accepted.
      operationId: patchRepository
      parameters:
        - $ref: '#/components/parameters/RepositoryID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fork]
              properties:
                fork:
                  type: string
                  description: Owner/name of the staging fork prepared for patch work.
      responses:
        '204':
          description: Repository updated
        '400':
          description: Body is not valid JSON
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422':
          description: No writable fields were supplied
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500':
          description: Database update failed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /repositories/{id}/scans:
    get:
      summary: List scans for a repository
      operationId: listRepositoryScans
      parameters:
        - $ref: '#/components/parameters/RepositoryID'
        - name: status
          in: query
          schema:
            type: string
            enum: [queued, running, paused, done, failed, skipped, cancelled]
        - name: skill
          in: query
          schema: { type: string }
          description: Filter by skill name (matches Scan.skill_name).
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Scan' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /repositories/{id}/maintainers:
    get:
      summary: List maintainers linked to this repository
      operationId: listRepositoryMaintainers
      parameters:
        - $ref: '#/components/parameters/RepositoryID'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Maintainer' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /repositories/{id}/packages:
    get:
      summary: List published packages produced by this repository
      operationId: listRepositoryPackages
      parameters:
        - $ref: '#/components/parameters/RepositoryID'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Package' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /repositories/{id}/alternatives:
    get:
      summary: List operator-curated package migration alternatives
      operationId: listRepositoryAlternatives
      parameters:
        - $ref: '#/components/parameters/RepositoryID'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PackageAlternative' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /repositories/{id}/advisories:
    get:
      summary: List published advisories affecting this repository's packages
      operationId: listRepositoryAdvisories
      parameters:
        - $ref: '#/components/parameters/RepositoryID'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Advisory' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /repositories/{id}/dependents:
    get:
      summary: List downstream dependents of this repository's packages
      operationId: listRepositoryDependents
      parameters:
        - $ref: '#/components/parameters/RepositoryID'
        - $ref: '#/components/parameters/RepositoryReadLimit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Dependent' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /repositories/{id}/ecosystems/{source}/raw:
    get:
      summary: Verbatim cached ecosyste.ms payload for one source
      description: >
        Returns the raw upstream payload pre-fetched and cached server-side.
        A 404 signals nothing is cached, in which case the caller falls
        back to a direct WebFetch. When the operator sets
        ecosystems_enrichment false, no source is ever cached, so every
        request answers 404 and the fallback is the only path.
      operationId: getRepositoryEcosystemsRaw
      parameters:
        - $ref: '#/components/parameters/RepositoryID'
        - name: source
          in: path
          required: true
          schema:
            type: string
            enum: [repo, packages, advisories, commits, issues, dependents]
      responses:
        '200':
          description: Cached payload, returned verbatim
          content:
            application/json: {}
        '400':
          description: Unknown source
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /repositories/{id}/expected:
    get:
      summary: List expected findings for benchmark comparison
      operationId: listRepositoryExpectedFindings
      parameters:
        - $ref: '#/components/parameters/RepositoryID'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ExpectedFinding' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /repositories/{id}/dependencies:
    get:
      summary: List dependencies declared by this repository's manifests
      operationId: listRepositoryDependencies
      parameters:
        - $ref: '#/components/parameters/RepositoryID'
        - $ref: '#/components/parameters/RepositoryReadLimit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Dependency' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /repositories/{id}/findings:
    get:
      summary: List findings across every scan on this repository
      operationId: listRepositoryFindings
      parameters:
        - $ref: '#/components/parameters/RepositoryID'
        - name: severity
          in: query
          schema: { type: string, enum: [Critical, High, Medium, Low] }
        - name: status
          in: query
          schema:
            type: string
            enum: [new, enriched, triaged, ready, reported, acknowledged, fixed, published, rejected, duplicate]
        - name: skill
          in: query
          description: Only return findings produced by scans of this skill (e.g. `semgrep`, `security-deep-dive`).
          schema: { type: string }
        - name: scan_group
          in: query
          description: Only return findings from scans in this parallel batch, so an in-flight audit skill reads only its siblings' findings.
          schema: { type: string }
        - name: sub_path
          in: query
          description: Only return findings for this monorepo sub-package (e.g. `activesupport`), so a repository can be exported or bundled one sub-package at a time.
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/FindingSummary' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      summary: Stream one finding into the concurrent-finding log
      description: >
        Records a single finding emitted mid-scan so sibling scans in the same
        `scan_group` can read it before this scan completes. The body is one
        finding in the same shape it has in the skill's `report.json`; the
        scan's identity (scan id, repository, commit, sub-path) is stamped from
        the bearer token, never trusted from the body. The finding still
        belongs in the final `report.json`, which reconciles against the
        streamed row by fingerprint rather than duplicating it.
      operationId: streamRepositoryFinding
      parameters:
        - $ref: '#/components/parameters/RepositoryID'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FindingInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FindingSummary' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /findings/{id}/notes:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer, format: int64 }
    get:
      summary: List notes on a finding
      operationId: listFindingNotes
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/FindingNote' }
    post:
      summary: Append a note to a finding
      operationId: addFindingNote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string }
                by:   { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FindingNote' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    RepositoryID:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
    RepositoryReadLimit:
      name: limit
      in: query
      description: Maximum number of rows to return; defaults to 200 and is capped at 1000.
      schema: { type: integer, minimum: 0, maximum: 1000, default: 200 }
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Token does not grant access to this resource
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...